update factory
Since the factory for Yelp API isn't running correctly, I'm temporarily
making the factory return hard coded data in it's place to act as the
Yelp factory.
When a user clicks the search button, markers are added to the map using
the factory data's coordinates, and the list of search results populate
the table below
Each marker can display a popup with the restaurant name and rating.