Added angular leaflet directive and node.js
Replaced the original leaflet api with anguler-leaflet-directive to make
it easier to incorporate leaflet and mapbox into the app.
added nodejs and started testing the app locally.
Yelp API is still giving an error when trying to connect.
The error is "No 'Access-Control-Allow-Origin' header is present on the
requested resource."
Searching online resources for a solution says this is due to the
browser settings restricting the connection. Need to somehow change
settings on local server to allow for CORS. I'm still not sure how this
works exactly so I hardcorded a mock JSON object to use while I continue
to code the rest of the functions for the app. I plan to return to this
issue after the other things are completed.