Updated markers on map
When the user clicks the search button, the yelp factory returns the
mock JSON data ($http still returns 'Access-Control-Allow-Origin'
error). the data populates the map with markers that contain the
restaurant's name and rating. The list of restaurants is also displayed
below the map with each restaurant's name, address, and rating, as well
as a delete button that will delete it from the list as well as removing
the marker from the map. There is also the option to remove all.